Move library components into the feature directory
