+++ src/apps/modern/features/libraries/components/NewCollectionButton.tsx
